+++ modules/ui-bootstrap/src/dynamic-form-bootstrap.component.spec.ts
@@ -6,7 +6,8 @@ import {
     DynamicFormsCoreModule,
     DynamicFormService,
     DynamicInputModel,
-    DynamicFormControlModel
+    DynamicFormControlModel,
+    CoreFormControlType
 } from "@ng2-dynamic-forms/core";
 import { DynamicFormBootstrapComponent, DYNAMIC_FORM_UI_BOOTSTRAP } from "./dynamic-form-bootstrap.component";
 
@@ -71,6 +72,8 @@ describe("DynamicFormBootstrapComponent test suite", () => {
         expect(component.isValid).toBe(true);
         expect(component.isInvalid).toBe(false);
         expect(component.showErrorMessages).toBe(false);
+
+        expect(component.formControlType).toBe(CoreFormControlType.Input);
     });
 
     it("should have an input element", () => {